refactor(commands): get projectName before branching off into different CLI workflows
