+++ .changeset/fast-fans-breathe.md
@@ -1,0 +1,5 @@
+---
+'create-expo-stack': patch
+---
+
+refactor(commands): get projectName before branching off into CLI
